Define basic structure of the entity interfaces

+++ src/main/java/com/bank/entity/Transaction.java
@@ -19,7 +19,7 @@ public class Transaction {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    public int transactionId;
+    public long transactionId;
 
     private int senderAccountNumber;
     private int receiverAccountNumber;
